Patches to address bugs and warnings due to turning on -Wall and -Werror on gcc3.4.3.  Also, Makefile.in modified to include -Wall and -Werror by default.  Details are at: http://mailman.isi.edu/pipermail/ns-developers/2005-July/001780.html
Contributed by Mathieu Lacage

+++ wpan/p802_15_4phy.cc
@@ -13,7 +13,7 @@
 // File:  p802_15_4phy.cc
 // Mode:  C++; c-basic-offset:8; tab-width:8; indent-tabs-mode:t
 
-// $Header: /home/smtatapudi/Thesis/nsnam/nsnam/ns-2/wpan/p802_15_4phy.cc,v 1.1 2005/01/24 18:34:25 haldar Exp $
+// $Header: /home/smtatapudi/Thesis/nsnam/nsnam/ns-2/wpan/p802_15_4phy.cc,v 1.2 2005/07/13 03:51:33 tomh Exp $
 
 /*
  * Copyright (c) 2003-2004 Samsung Advanced Institute of Technology and
@@ -556,13 +556,13 @@ void Phy802_15_4::recv(Packet *p, Handler *h)
 			rxEDPeakPower = rxTotPower[ppib.phyCurrentChannel];
 		
 		if ((p802_15_4macDA(p) == index_)			//packet for this node
-		  ||(p802_15_4macDA(p) == MAC_BROADCAST))		//broadcast packet
+		  ||(p802_15_4macDA(p) == ((int)MAC_BROADCAST)))		//broadcast packet
 			rxThisTotNum[wph->phyCurrentChannel]++;
 
 		if (trx_state == p_RX_ON)
 		if (wph->phyCurrentChannel == ppib.phyCurrentChannel)	//current channel
 		if ((p802_15_4macDA(p) == index_)			//packet for this node
-		  ||(p802_15_4macDA(p) == MAC_BROADCAST))		//broadcast packet
+		  ||(p802_15_4macDA(p) == ((int)MAC_BROADCAST)))		//broadcast packet
 		if (wph->SHR_SFD == defSHR_SFD)				//valid SFD
 		{
 #ifdef DEBUG802_15_4
@@ -687,7 +687,7 @@ void Phy802_15_4::recvOverHandler(Packet *p)
 		rxTotPower[wph->phyCurrentChannel] = 0.0;
 
 	if ((p802_15_4macDA(p) != index_)
-	  &&(p802_15_4macDA(p) != MAC_BROADCAST))	//packet not for this node (interference)
+	  &&(p802_15_4macDA(p) != ((int)MAC_BROADCAST)))	//packet not for this node (interference)
 		Packet::free(p);
 	else if (p != rxPkt)				//packet corrupted (not the strongest one) or un-detectable
 	{
